Refactor : Firebase Util 파일 위치 변경
