+++ src/main/java/com/mentionall/cpr2u/call/service/CprCallService.java
@@ -11,11 +11,12 @@ import com.mentionall.cpr2u.user.domain.AngelStatus;
 import com.mentionall.cpr2u.user.domain.User;
 import com.mentionall.cpr2u.user.repository.address.AddressRepository;
 import com.mentionall.cpr2u.user.repository.device_token.DeviceTokenRepository;
-import com.mentionall.cpr2u.util.fcm.FcmMessage;
 import com.mentionall.cpr2u.util.exception.CustomException;
 import com.mentionall.cpr2u.util.exception.ResponseCode;
 import com.mentionall.cpr2u.util.fcm.FcmDataType;
+import com.mentionall.cpr2u.util.fcm.FcmMessage;
 import com.mentionall.cpr2u.util.fcm.FcmPushType;
+import com.mentionall.cpr2u.util.fcm.FirebaseCloudMessageUtil;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.PageRequest;
