+++ src/test/java/com/mentionall/cpr2u/TestConfig.java
@@ -1,7 +1,7 @@
 package com.mentionall.cpr2u;
 
 import com.mentionall.cpr2u.call.service.FakeFirebaseCloudMessageUtil;
-import com.mentionall.cpr2u.call.service.FirebaseCloudMessageUtil;
+import com.mentionall.cpr2u.util.fcm.FirebaseCloudMessageUtil;
 import com.mentionall.cpr2u.util.twilio.FakeTwilioUtil;
 import com.mentionall.cpr2u.util.twilio.TwilioUtil;
 import org.springframework.context.annotation.Bean;
