+++ src/test/java/com/mentionall/cpr2u/call/service/FakeFirebaseCloudMessageUtil.java
@@ -1,5 +1,7 @@
 package com.mentionall.cpr2u.call.service;
 
+import com.mentionall.cpr2u.util.fcm.FirebaseCloudMessageUtil;
+
 import java.util.List;
 import java.util.Map;
